+++ parrot/src/color.rs
@@ -91,6 +91,10 @@ impl Rgba {
 
     pub const RED: Self = Self::new(1.0, 0.0, 0.0, 1.0);
 
+    pub const WHITE: Self = Self::new(1.0, 1.0, 1.0, 1.0);
+
+    pub const BLACK: Self = Self::new(0.0, 0.0, 0.0, 1.0);
+
     /// Create a new colour with corresponding values between 0 and 1
     pub const fn new(r: f32, g: f32, b: f32, a: f32) -> Self {
         Self {