+++ parrot/src/transform.rs
@@ -3,3 +3,6 @@ pub struct ScreenSpace;
 
 /// Worldspace unit tag for euclid
 pub struct WorldSpace;
+
+/// Objectspace unit tag for euclid
+pub struct ObjectSpace;