+++ pigeon/src/lib.rs
@@ -1,0 +1,8 @@
+extern crate pigeon_parrot as parrot;
+
+pub mod pipeline;
+pub mod graphics;
+pub mod pigeon;
+
+pub use pigeon::Pigeon;
+pub use parrot::transform;